feat:REQ-050426 agregar presentaciones para registro de venta y compra

+++ src/main/java/com/svc/ventas/models/mapstruct/dto/ProductoMasVendidoDTO.java
@@ -14,11 +14,9 @@ public class ProductoMasVendidoDTO implements Serializable {
 
   private String nombre;
 
-  private Long cantidadVendida;
+  private BigDecimal cantidadVendida;
 
   private BigDecimal valorVentas;
-
-  //private Integer stock;
 
   private Boolean estado;
